requirements.txt
requests
pyyaml
tenacity
tqdm
